Checkpoint. Parse in DOSBOX is works. (AZTEK 4.2)

+++ types.h
@@ -27,8 +27,8 @@ typedef unsigned char uint8_t;
 typedef signed char int8_t;
 typedef unsigned short uint16_t;
 typedef signed short int16_t;
-typedef unsigned int uint32_t;
-typedef signed int int32_t;
+typedef unsigned long int uint32_t;
+typedef signed long int int32_t;
 /*typedef unsigned long long u64;
 typedef long long i64;*/
 #ifndef TRUE
